Deal with rate limit
You can retrieve the rate Limit that GitHub allows you with the code of this PR.  

I calculated the rate value necessary to download all ThinkR repositories, according to calculation of GitHub rate limit minutes.  
I saw that I needed at least a score of 1250 before starting to fetch repo. 
If you wait for the next hour when the rate is low, then the CI won't fail.